fixes /config and cleanup tests
