+++ services/web/server/src/simcore_service_webserver/rest_handlers.py
@@ -13,6 +13,7 @@ from .application_settings import APP_SETTINGS_KEY
 from .products import get_product_name
 from .redis import get_redis_scheduled_maintenance_client
 from .rest_healthcheck import HealthCheck, HealthCheckFailed
+from .utils_aiohttp import envelope_json_response
 
 log = logging.getLogger(__name__)
 
@@ -55,8 +56,7 @@ async def healthcheck_readiness_probe(request: web.Request):
     health_report = healthcheck.get_app_info(request.app)
     # NOTE: do NOT run healthcheck here, just return info fast.
     health_report["status"] = "SERVICE_RUNNING"
-
-    return web.json_response(data={"data": health_report})
+    return envelope_json_response(health_report)
 
 
 @routes.get(f"/{API_VTAG}/config", name="get_config")
@@ -78,7 +78,7 @@ async def get_config(request: web.Request):
         product_name, {}
     )
 
-    return web.json_response(data={"data": app_public_config | product_public_config})
+    return envelope_json_response(app_public_config | product_public_config)
 
 
 @routes.get(f"/{API_VTAG}/scheduled_maintenance", name="get_scheduled_maintenance")
@@ -90,4 +90,5 @@ async def get_scheduled_maintenance(request: web.Request):
     # {"start": "2023-01-17T14:45:00.000Z", "end": "2023-01-17T23:00:00.000Z", "reason": "Release 1.0.4"}
     if maintenance_str := await redis_client.get(hash_key):
         return web.json_response(data={"data": maintenance_str})
+
     return web.json_response(status=204)
